data-pipelines-airflow/requirements.txt
alembic==1.7.4
anyio==3.3.4
apache-airflow==2.2.2
apache-airflow-providers-amazon==2.4.0
apache-airflow-providers-ftp==2.0.1
apache-airflow-providers-http==2.0.1
apache-airflow-providers-imap==2.0.1
apache-airflow-providers-postgres==2.3.0
apache-airflow-providers-sqlite==2.0.1
apispec==3.3.2
argcomplete==1.12.3
asn1crypto==1.4.0
attrs==20.3.0
Babel==2.9.1
bcrypt==3.2.0
beautifulsoup4==4.10.0
blinker==1.4
boto3==1.18.65
botocore==1.21.65
cached-property==1.5.2
cattrs==1.5.0
certifi==2020.12.5
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
clickclick==20.10.2
colorama==0.4.4
colorlog==4.8.0
commonmark==0.9.1
configparser==3.5.3
croniter==1.0.15
cryptography==3.4.8
decorator==5.1.0
defusedxml==0.7.1
dill==0.3.1.1
distro==1.6.0
dnspython==2.1.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.16
email-validator==1.1.3
Flask==1.1.2
Flask-Admin==1.5.4
Flask-AppBuilder==3.3.4
Flask-Babel==2.0.0
Flask-Caching==1.10.1
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.3.0
Flask-SQLAlchemy==2.5.1
flask-swagger==0.2.13
Flask-WTF==0.14.3
funcsigs==1.0.2
future==0.16.0
graphviz==0.18
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
httpcore==0.13.7
httpx==0.20.0
idna==3.3
importlib-metadata==4.8.1
importlib-resources==5.4.0
inflection==0.5.1
iso8601==1.0.0
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
json-merge-patch==0.2
jsonpath-ng==1.5.3
jsonschema==3.2.0
lazy-object-proxy==1.4.3
lockfile==0.12.2
lxml==4.6.4
Mako==1.1.5
Markdown==3.3.4
MarkupSafe==1.1.1
marshmallow==3.14.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
natsort==8.0.0
numpy==1.21.4
openapi-schema-validator==0.1.5
openapi-spec-validator==0.3.1
packaging==21.0
pandas==0.25.3
paramiko==2.8.0
pendulum==2.1.2
ply==3.11
prison==0.2.1
psutil==5.8.0
psycopg2-binary==2.9.2
pycparser==2.21
Pygments==2.10.0
PyJWT==1.7.1
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-daemon==2.3.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2021.3
pytzdata==2020.1
PyYAML==5.4.1
redshift-connector==2.0.900
requests==2.26.0
rfc3986==1.5.0
rich==10.13.0
s3transfer==0.5.0
scramp==1.4.1
setproctitle==1.2.2
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
SQLAlchemy==1.3.24
SQLAlchemy-JSONField==1.0.0
sqlalchemy-redshift==0.8.8
SQLAlchemy-Utils==0.37.9
swagger-ui-bundle==0.0.9
tabulate==0.8.9
tenacity==8.0.1
termcolor==1.1.0
text-unidecode==1.3
texttable==1.6.4
thrift==0.15.0
typing-extensions==3.10.0.2
tzlocal==1.5.1
unicodecsv==0.14.1
urllib3==1.26.7
watchtower==1.0.6
websocket-client==0.59.0
Werkzeug==1.0.1
WTForms==2.3.3
zipp==3.6.0
zope.deprecation==4.4.0
